sparql_validator/requirements.txt
python-decouple==3.8
SPARQLWrapper==2.0.0
pytest==7.4.3